+++ app/Http/Livewire/Admin/Categorypage/Electronic/Slider.php
@@ -82,8 +82,9 @@ class Slider extends Component
 //        $slider = DB::connection('mysql-electronic')->table('category_electronic_slider')
         $slider = DB::table('category_electronic_slider')
             ->where('id',$id)->limit($id);
-        Storage::disk('public')->delete("storage", $slider2->img);
-        $slider->delete();
+        if ($slider->img) {
+            Storage::disk('public')->delete("storage", $slider2->img);
+        }$slider->delete();
 
         Log::create([
             'user_id' => auth()->user()->id,
